# exec: use go runtime hash functions in hashjoiner