--- a/pkg/sql/exec/noescape.s
+++ b/pkg/sql/exec/noescape.s
@@ -1,0 +1,1 @@
+// Empty assembly file to allow go:linkname to work.