@asubiotto 2019-03-01
Testing this on a sql-20 hash join at scalefactor 1 results in an average `1.8%` increase in latency. However, the results varied by as much as `3.5%` between each run.