@asubiotto 2019-03-04
TFTRs

@yahor, I'm relying on the smhasher "quality" results of https://github.com/Cyan4973/xxHash (the functions are apparently derived from xxhash and cityhash).

bors r=yuzefovich,jordanlewis